finish

+++ VideoRentalWeb/VideoRentalWeb/Controllers/ClienteleController.cs
@@ -23,49 +23,55 @@ public class ClienteleController : Controller
 
     private const string FilterKey = "clientele";
 
-    public ClienteleController( UserManager<User> userManager, VideoRentalContext context, CacheProvider cacheProvider)
+    public ClienteleController(UserManager<User> userManager, VideoRentalContext context, CacheProvider cacheProvider)
     {
         _db = context;
         _cache = cacheProvider;
-        //_roleManager = roleManager;
         _userManager = userManager;
     }
 
     public IActionResult Index(SortState sortState = SortState.ClientNameAsc, int page = 1)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            ClienteleFilterViewModel filter = HttpContext.Session.Get<ClienteleFilterViewModel>(FilterKey);
-            if (filter == null)
+            var currentUser = _userManager.GetUserAsync(User).Result;
+
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
             {
-                filter = new ClienteleFilterViewModel() { Name = string.Empty };
-                HttpContext.Session.Set(FilterKey, filter);
+                ClienteleFilterViewModel filter = HttpContext.Session.Get<ClienteleFilterViewModel>(FilterKey);
+                if (filter == null)
+                {
+                    filter = new ClienteleFilterViewModel() { Name = string.Empty };
+                    HttpContext.Session.Set(FilterKey, filter);
+                }
+
+                string modelKey = $"{typeof(Clientele).Name}-{page}-{sortState}-{filter.Name}";
+                if (!_cache.TryGetValue(modelKey, out ClienteleViewModel model))
+                {
+                    model = new ClienteleViewModel();
+
+                    IQueryable<Clientele> clienteles = GetSortedEntities(sortState, filter.Name);
+
+                    int count = clienteles.Count();
+                    int pageSize = 10;
+                    model.PageViewModel = new PageViewModel(page, count, pageSize);
+
+                    model.Entities = count == 0 ? new List<Clientele>() : clienteles.Skip((model.PageViewModel.CurrentPage - 1) * pageSize).Take(pageSize).ToList();
+                    model.SortViewModel = new SortViewModel(sortState);
+                    model.ClienteleFilterViewModel = filter;
+
+                    _cache.Set(modelKey, model);
+                }
+
+                return View(model);
             }
 
-            string modelKey = $"{typeof(Clientele).Name}-{page}-{sortState}-{filter.Name}";
-            if (!_cache.TryGetValue(modelKey, out ClienteleViewModel model))
+            else
             {
-                model = new ClienteleViewModel();
-
-                IQueryable<Clientele> clienteles = GetSortedEntities(sortState, filter.Name);
-
-                int count = clienteles.Count();
-                int pageSize = 10;
-                model.PageViewModel = new PageViewModel(page, count, pageSize);
-
-                model.Entities = count == 0 ? new List<Clientele>() : clienteles.Skip((model.PageViewModel.CurrentPage - 1) * pageSize).Take(pageSize).ToList();
-                model.SortViewModel = new SortViewModel(sortState);
-                model.ClienteleFilterViewModel = filter;
-
-                _cache.Set(modelKey, model);
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(model);
         }
-
         else
         {
             return RedirectToAction("Index", "Home");
@@ -75,21 +81,28 @@ public class ClienteleController : Controller
     [HttpPost]
     public IActionResult Index(ClienteleFilterViewModel filterModel, int page)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            ClienteleFilterViewModel filter = HttpContext.Session.Get<ClienteleFilterViewModel>(FilterKey);
-            if (filter != null)
+            var currentUser = _userManager.GetUserAsync(User).Result;
+
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
             {
-                filter.Name = filterModel.Name;
+                ClienteleFilterViewModel filter = HttpContext.Session.Get<ClienteleFilterViewModel>(FilterKey);
+                if (filter != null)
+                {
+                    filter.Name = filterModel.Name;
 
-                HttpContext.Session.Remove(FilterKey);
-                HttpContext.Session.Set(FilterKey, filter);
+                    HttpContext.Session.Remove(FilterKey);
+                    HttpContext.Session.Set(FilterKey, filter);
+                }
+
+                return RedirectToAction("Index", new { page });
             }
-
-            return RedirectToAction("Index", new { page });
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         else
         {
@@ -99,17 +112,24 @@ public class ClienteleController : Controller
 
     public IActionResult Create(int page)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            ClienteleViewModel model = new ClienteleViewModel()
-            {
-                PageViewModel = new PageViewModel { CurrentPage = page }
-            };
+            var currentUser = _userManager.GetUserAsync(User).Result;
 
-            return View(model);
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+            {
+                ClienteleViewModel model = new ClienteleViewModel()
+                {
+                    PageViewModel = new PageViewModel { CurrentPage = page }
+                };
+
+                return View(model);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         else
         {
@@ -120,33 +140,40 @@ public class ClienteleController : Controller
     [HttpPost]
     public async Task<IActionResult> Create(ClienteleViewModel model)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            foreach (var entry in ModelState)
-            {
-                var key = entry.Key; // Название свойства
-                var errors = entry.Value.Errors.Select(e => e.ErrorMessage).ToList(); // Список ошибок для свойства
+            var currentUser = _userManager.GetUserAsync(User).Result;
 
-                // Далее можно использовать key и errors в соответствии с вашими потребностями
-                Console.WriteLine($"Property: {key}, Errors: {string.Join(", ", errors)}");
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+            {
+                foreach (var entry in ModelState)
+                {
+                    var key = entry.Key; // Название свойства
+                    var errors = entry.Value.Errors.Select(e => e.ErrorMessage).ToList(); // Список ошибок для свойства
+
+                    // Далее можно использовать key и errors в соответствии с вашими потребностями
+                    Console.WriteLine($"Property: {key}, Errors: {string.Join(", ", errors)}");
+                }
+
+                if (ModelState.IsValid)
+                {
+                    await _db.Clienteles.AddAsync(model.Entity);
+                    await _db.SaveChangesAsync();
+
+                    _cache.Clean();
+
+                    return RedirectToAction("Index", "Clientele");
+                }
+
+                return View(model);
             }
 
-            if (ModelState.IsValid)
+            else
             {
-                await _db.Clienteles.AddAsync(model.Entity);
-                await _db.SaveChangesAsync();
-
-                _cache.Clean();
-
-                return RedirectToAction("Index", "Clientele");
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(model);
         }
-
         else
         {
             return RedirectToAction("Index", "Home");
@@ -155,21 +182,28 @@ public class ClienteleController : Controller
 
     public async Task<IActionResult> Edit(int id, int page)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            Clientele clientele = await _db.Clienteles.FindAsync(id);
-            if (clientele != null)
-            {
-                ClienteleViewModel model = new ClienteleViewModel();
-                model.PageViewModel = new PageViewModel { CurrentPage = page };
-                model.Entity = clientele;
+            var currentUser = _userManager.GetUserAsync(User).Result;
 
-                return View(model);
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+            {
+                Clientele clientele = await _db.Clienteles.FindAsync(id);
+                if (clientele != null)
+                {
+                    ClienteleViewModel model = new ClienteleViewModel();
+                    model.PageViewModel = new PageViewModel { CurrentPage = page };
+                    model.Entity = clientele;
+
+                    return View(model);
+                }
+                return NotFound();
             }
-            return NotFound();
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         else
         {
@@ -180,37 +214,44 @@ public class ClienteleController : Controller
     [HttpPost]
     public async Task<IActionResult> Edit(ClienteleViewModel model)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            if (ModelState.IsValid)
+            var currentUser = _userManager.GetUserAsync(User).Result;
+
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
             {
-                Clientele clientele = _db.Clienteles.Find(model.Entity.ClientId);
-                if (clientele != null)
+                if (ModelState.IsValid)
                 {
-                    clientele.Name = model.Entity.Name;
-                    clientele.Surname = model.Entity.Surname;
-                    clientele.Middlename = model.Entity.Middlename;
-                    clientele.Phone = model.Entity.Phone;
-                    clientele.Passport = model.Entity.Passport;
-                    clientele.Addres = model.Entity.Addres;
+                    Clientele clientele = _db.Clienteles.Find(model.Entity.ClientId);
+                    if (clientele != null)
+                    {
+                        clientele.Name = model.Entity.Name;
+                        clientele.Surname = model.Entity.Surname;
+                        clientele.Middlename = model.Entity.Middlename;
+                        clientele.Phone = model.Entity.Phone;
+                        clientele.Passport = model.Entity.Passport;
+                        clientele.Addres = model.Entity.Addres;
 
-                    _db.Clienteles.Update(clientele);
-                    await _db.SaveChangesAsync();
+                        _db.Clienteles.Update(clientele);
+                        await _db.SaveChangesAsync();
 
-                    _cache.Clean();
+                        _cache.Clean();
 
-                    return RedirectToAction("Index", "Clientele", new { page = model.PageViewModel.CurrentPage });
+                        return RedirectToAction("Index", "Clientele", new { page = model.PageViewModel.CurrentPage });
+                    }
+                    else
+                    {
+                        return NotFound();
+                    }
                 }
-                else
-                {
-                    return NotFound();
-                }
+
+                return View(model);
             }
-
-            return View(model);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         else
         {
@@ -220,24 +261,31 @@ public class ClienteleController : Controller
 
     public async Task<IActionResult> Delete(int id, int page)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            Clientele clientele = await _db.Clienteles.FindAsync(id);
-            if (clientele == null)
-                return NotFound();
+            var currentUser = _userManager.GetUserAsync(User).Result;
 
-            bool deleteFlag = false;
-            string message = "Do you want to delete this entity";
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+            {
+                Clientele clientele = await _db.Clienteles.FindAsync(id);
+                if (clientele == null)
+                    return NotFound();
 
-            ClienteleViewModel model = new ClienteleViewModel();
-            model.Entity = clientele;
-            model.PageViewModel = new PageViewModel { CurrentPage = page };
-            model.DeleteViewModel = new DeleteViewModel { Message = message, IsDeleted = deleteFlag };
+                bool deleteFlag = false;
+                string message = "Do you want to delete this entity";
 
-            return View(model);
+                ClienteleViewModel model = new ClienteleViewModel();
+                model.Entity = clientele;
+                model.PageViewModel = new PageViewModel { CurrentPage = page };
+                model.DeleteViewModel = new DeleteViewModel { Message = message, IsDeleted = deleteFlag };
+
+                return View(model);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         else
         {
@@ -248,23 +296,30 @@ public class ClienteleController : Controller
     [HttpPost]
     public async Task<IActionResult> Delete(ClienteleViewModel model)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            Clientele clientele = await _db.Clienteles.FindAsync(model.Entity.ClientId);
-            if (clientele == null)
-                return NotFound();
+            var currentUser = _userManager.GetUserAsync(User).Result;
 
-            _db.Clienteles.Remove(clientele);
-            await _db.SaveChangesAsync();
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+            {
+                Clientele clientele = await _db.Clienteles.FindAsync(model.Entity.ClientId);
+                if (clientele == null)
+                    return NotFound();
 
-            _cache.Clean();
+                _db.Clienteles.Remove(clientele);
+                await _db.SaveChangesAsync();
 
-            model.DeleteViewModel = new DeleteViewModel { Message = "The entity was successfully deleted.", IsDeleted = true };
+                _cache.Clean();
 
-            return View(model);
+                model.DeleteViewModel = new DeleteViewModel { Message = "The entity was successfully deleted.", IsDeleted = true };
+
+                return View(model);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
         else
         {
@@ -274,27 +329,34 @@ public class ClienteleController : Controller
 
     public async Task<IActionResult> Details(int id)
     {
-        var currentUser = _userManager.GetUserAsync(User).Result;
-
-        // Проверка наличия роли Admin у текущего пользователя
-        if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
+        if (User.Identity.IsAuthenticated)
         {
-            Clientele clientele = await _db.Clienteles.FindAsync(id);
-            if (clientele != null)
+            var currentUser = _userManager.GetUserAsync(User).Result;
+
+            // Проверка наличия роли Admin у текущего пользователя
+            if (_userManager.IsInRoleAsync(currentUser, "Admin").Result || _userManager.IsInRoleAsync(currentUser, "Manager").Result)
             {
-                ClienteleViewModel model = new ClienteleViewModel()
+                Clientele clientele = await _db.Clienteles.FindAsync(id);
+                if (clientele != null)
                 {
-                    Entity = clientele,
+                    ClienteleViewModel model = new ClienteleViewModel()
+                    {
+                        Entity = clientele,
 
-                    PageViewModel = new PageViewModel()
-                };
+                        PageViewModel = new PageViewModel()
+                    };
 
-                return View(model);
+                    return View(model);
+                }
+
+                return NotFound();
             }
 
-            return NotFound();
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
-
         else
         {
             return RedirectToAction("Index", "Home");
